move Portal back to lab
