+++ packages/lab/src/combo-box-deprecated/internal/DefaultComboBox.tsx
@@ -1,7 +1,5 @@
 import { flip, limitShift, shift, size } from "@floating-ui/react";
 import {
-  Portal,
-  Tooltip,
   useAriaAnnouncer,
   useFloatingUI,
   useForkRef,
@@ -17,6 +15,7 @@ import {
   useRef,
   useState,
 } from "react";
+import { Portal } from '../../portal'
 import {
   IndexedListItemProps,
   ListBase,
@@ -33,10 +32,10 @@ import { Input, InputProps } from "../../input";
 export type BaseComboBoxProps<
   Item,
   Variant extends ListSelectionVariant = "default"
-> = Omit<
-  HTMLAttributes<HTMLDivElement>,
-  "children" | "onChange" | "onSelect" | "onFocus" | "onBlur" | "onClick"
-> &
+  > = Omit<
+    HTMLAttributes<HTMLDivElement>,
+    "children" | "onChange" | "onSelect" | "onFocus" | "onBlur" | "onClick"
+  > &
   Pick<
     ListProps<Item, Variant>,
     | "displayedItemCount"
@@ -67,7 +66,7 @@ export type BaseComboBoxProps<
 
 export interface DefaultComboBoxProps<Item>
   extends BaseComboBoxProps<Item>,
-    Pick<InputProps, "onFocus" | "onBlur"> {
+  Pick<InputProps, "onFocus" | "onBlur"> {
   InputProps?: InputProps;
   initialSelectedItem?: Item;
   selectedItem?: Item;
@@ -132,16 +131,16 @@ export const DefaultComboBox = function DefaultComboBox<Item>(
   const middleware = isDesktop
     ? []
     : [
-        flip({
-          fallbackPlacements: ["bottom-start", "top-start"],
-        }),
-        shift({ limiter: limitShift() }),
-        size({
-          apply({ availableHeight }) {
-            setMaxListHeight(availableHeight);
-          },
-        }),
-      ];
+      flip({
+        fallbackPlacements: ["bottom-start", "top-start"],
+      }),
+      shift({ limiter: limitShift() }),
+      size({
+        apply({ availableHeight }) {
+          setMaxListHeight(availableHeight);
+        },
+      }),
+    ];
   const { reference, floating, x, y, strategy } = useFloatingUI({
     placement: "bottom-start",
     middleware,
