+++ packages/lab/src/query-input/internal/ValueSelector.tsx
@@ -1,6 +1,5 @@
 import {
   makePrefixer,
-  Portal,
   useFloatingUI,
   useForkRef,
   useIsomorphicLayoutEffect,
@@ -10,6 +9,7 @@ import { QueryInputCategory, QueryInputItem } from "../queryInputTypes";
 import { CategoryList } from "./CategoryList";
 import { SearchList } from "./SearchList";
 import { ValueList } from "./ValueList";
+import { Portal } from '../../portal'
 import { useWindow } from "../../window";
 
 const withBaseName = makePrefixer("saltQueryInputValueSelector");
