+++ packages/lab/src/tabs/drag-drop/Draggable.tsx
@@ -1,6 +1,7 @@
-import { Portal, useForkRef } from "@salt-ds/core";
+import { useForkRef } from "@salt-ds/core";
 import { clsx } from "clsx";
 import { forwardRef, MutableRefObject, useCallback } from "react";
+import { Portal } from '../../portal'
 import { Rect } from "./dragDropTypes";
 
 import "./Draggable.css";
